+++ scriptabit/task_service.py
@@ -29,16 +29,6 @@ class TaskService(object):
         """
         raise NotImplementedError
 
-    # def get_task(self, _id):
-        # """ Gets a task by id.
-
-        # Args:
-            # _id (str): The task ID to get.
-
-        # Returns: Task: The task if it exists, otherwise None.
-        # """
-        # raise NotImplementedError
-
     def persist_tasks(self, tasks):
         """ Persists the tasks.